+++ src/codecs/index.ts
@@ -1,2 +1,3 @@
 export * from './docx'
 
+export * from './pdf'